requirements.txt
numpy==1.24.3
gymnasium==0.29.1
stable-baselines3==2.1.0
selenium==4.15.2
opencv-python==4.8.1.78
torch==2.1.0
streamlit==1.28.1
imageio==2.31.6
webdriver-manager==4.0.1
playwright==1.43.0
colorama==0.4.6
tensorboard==2.15.1
matplotlib==3.8.2
pillow==10.1.0
